code refactoring (frontend) (wip)
