+++ Cargo.toml
@@ -4,6 +4,9 @@ version = "0.2.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+path = "src/app.rs"
+
 [dependencies]
 chrono = "0.4"
 console_error_panic_hook = "0.1"
