+++ src/main.rs
@@ -1,13 +1,11 @@
 // leptos_i18n::load_locales!();
 
-mod app;
-
 use leptos::prelude::mount_to_body;
 use leptos::*;
 use thaw::{ConfigProvider, ToasterProvider};
 
-use app::*;
 // use i18n::*;
+use md_minutes_generator_ui::App;
 
 fn main() {
     console_error_panic_hook::set_once();
